Extract iroh helpers
